Move CV pipeline into its own file so that it's easier to edit multiple pipelines
